Don't try to delete file if not exist (throws exception on Windows).

+++ src/test/java/com/github/alexdlaird/ngrok/process/NgrokProcessTest.java
@@ -179,7 +179,9 @@ public class NgrokProcessTest extends NgrokTestCase {
     @Test
     public void testStartNoBinary() throws IOException {
         // GIVEN
-        Files.delete(javaNgrokConfig.getNgrokPath());
+        if (Files.exists(javaNgrokConfig.getNgrokPath())) {
+            Files.delete(javaNgrokConfig.getNgrokPath());
+        }
 
         // WHEN
         final NgrokException exception = assertThrows(NgrokException.class, ngrokProcess::start);
